Handle bintray 409
When publishing InstallSpinnaker.sh to bintray from dev/build_spinnaker.sh, it
is not uncommon to get a 409 error. I think this is because the content changes but
version numbers do not (until the project is explicitly released). To work around
this, when encountering a 409 publising an update, try to delete the exiting content
and retry.

This needs more investigation, but this workaround is sufficient for now.
It is off by default so needs to be explicitly requested with a commandline arg.